Add files via upload
In this project, we wrote a python code from the scratch by just using the NumPy library to perform the Naïve Bayesian Classification method on a dataset provided along with the code. First, we simply made an excel file for the training dataset named “training_data.xlsx” (includes 20 samples of 10 private and 10 public) and an excel file for the testing dataset named “testing_data.xlsx” (includes 5 samples of 3 private and 2 public). Then, we used the pandas in python only for importing these two datasets in our code (pandas is much better than the numpy for loading a dataset in python).